requirements/base.txt
Django==3.0.14
wagtail==2.9.3
wagtailtrans==2.1
gunicorn==20.0.4
django-dotenv==1.4.2
django-robots==4.0
psycopg2-binary==2.8.5
xmljson==0.2.1
django-prettyjson==0.4.1
python-dateutil==2.8.1
marshmallow==3.12.1
django-debug-toolbar==2.2.1
pyyaml==5.4
retrie==0.1.2
django-cors-headers==3.7.0
nltk==3.6.2